# Replace `std::sync` locks with `parking_lot`