@Jake-Shadle 2019-09-17
Ahh yes, so I should have mentioned I did try this last week, but comparing the benchmarks showed it made more of them worse than better, unfortunately.