@repi 2019-09-17
Those results are a bit surprising.

Though at least nice to get rid of the unwraps from the mutex poisioning "handling".

Would be interesting to try on the work 10-core Intel desktops. Low-end laptop CPU perf is likely not super important but highest end AMD is also not that common, but workstation Intel CPU with 8-10 cores could be a good target