@h3r2tic 2019-09-18
Added timings from my work Xeon machine.