@h3r2tic 2019-09-18
Yep. I'll close this for now, but keep the branch in case we want to experiment a bit more later.